\name{fabric}
\alias{fabric}
\docType{data}
\title{Fabric Data
}
\description{
In the fabric data (Bissell, 1972), the main response is the number of faults
in a bolt of fabric of length.
}
\usage{data("fabric")}
\format{
  A data frame with 32 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{log of length}
    \item{\code{y}}{number of faults}
    \item{\code{rf}}{obervation number for 32 obervations}
  }
}
\references{
Bissell, A.F. (1972). A negative binomial model with varying element
sizes. Biometrika, 59, 435--441.
}

