\name{factor}
\alias{factor}
\docType{data}
\title{Binary Factor Data}
\description{
Lee, Nelder and Pawitan considered the binary factor data with six items
taking 1 or 0 for 350 subjects.
}
\usage{data("factor")}
\format{
  A data frame with 350 observations on the following 7 variables.
  \describe{
    \item{\code{y1}}{item 1}
    \item{\code{y2}}{item 2}
    \item{\code{y3}}{item 3}
    \item{\code{y4}}{item 4}
    \item{\code{y5}}{item 5}
    \item{\code{y6}}{item 6}
    \item{\code{subject}}{subjects idenfier}
  }
}
\references{
Lee, Y., Nelder, J.A. and Pawitan, Y. (2016). Generalized Linear Models
with Random Effects (2nd Edition). Chapman and Hall, Boca Raton, 2nd edition.
}

