\name{gas}
\alias{gas}
\docType{data}
\title{Gas Consumption Data}
\description{
Durbin and Koopman (2000) considered the lagged quarterly-demand for
gas in the UK from 1960 to 1986.
}
\usage{data("gas")}
\format{
  A data frame with 108 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{lagged quarterly-demand for gas}
    \item{\code{year}}{year; 1960...1986}
    \item{\code{quarter}}{quarter; 1...4}
    \item{\code{time}}{data number for 108 obervations}
  }
}
\references{
Durbin, J. and Koopman, S.J. (2000). Time series analysis of nongaussian
observations based on state space models from both classical
and bayesian perspectives (with discussion). Journal of the Royal
Statistical Society: Series B (Statistical Methodology), 62, 3--56.
}

