\name{mdhglm-package}
\alias{mdhglm-package}
\docType{package}
\title{Multivariate Double Hierarchical Generalized Linear Models}
\description{
This package allows different models for multivariate response variables where each response is assumed to follow
DHGLMs (Lee and Nelder, 2006; Lee, Nelder and Pawitan, 2006, 2016). With DHGLMs, the mean and dispersion parameters for the variance of random effects and residual variance (overdispersion) can
be modeled as random-effect models. Furthermore, the mdhglm package allows various correlation structures among the random effects in
DHGLMs of different response variables. Functions are provided to format and summarize the mdhglm results. The estimates of fixed effects,
random effects and their variances and correlations are calculated.
}

\details{
\tabular{ll}{
Package: \tab mdhglm\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2016-09-19\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
This is version 1.4 of the mdhglm package.
}


\author{Youngjo Lee, Marek Molas, Maengseok Noh

Maintainer: Maengseok Noh <msnoh@pknu.ac.kr>
}

\references{
Lee, Y. and Nelder, J.A. (2006). Double hierarchical generalized linear models (with discussion), Applied Statistics 55, 139--185.

Lee, Y., Nelder, J.A. and Pawitan, Y. (2006). Generalized Linear Models
with Random Effects. Chapman and Hall, Boca Raton, 1st edition.

Lee, Y., Nelder, J.A. and Pawitan, Y. (2016). Generalized Linear Models
with Random Effects (2nd Edition). Chapman and Hall, Boca Raton,
2nd edition.

}
