\name{Orthodont}
\alias{Orthodont}
\docType{data}
\title{
Orthodontic Growth Data
}
\description{
The orthodontic growth data (Pinheiro and Bates, 2000) contain the
growth measurements of 27 children for 11 girls and 16 boys from age 8
until age 14. Every two years, the distance between the pituitary and the
pterygomaxillary fissure was recorded at ages 8, 10, 12 and 14.
}
\usage{data("Orthodont")}
\format{
  A data frame with 108 observations on the following 4 variables.
  \describe{
    \item{\code{distance}}{distance between the pituitary and the pterygomaxillary fissure}
    \item{\code{age}}{age in year (8, 10, 12 and 14 for each subject)}
    \item{\code{Subject}}{subjects indenfier for 11 girs and 16 boys}
    \item{\code{Sex}}{Gender (\code{Male} and \code{Female})}
  }
}
\references{
Pinheiro, J.C. and Bates, D. (2000). Mixed-Effects Models in S and
S-PLUS. Springer, New York.
}

