\name{Prestige}
\alias{Prestige}
\docType{data}
\title{Prestige Data
}
\description{
Fox, et al. (2016) considered an additive non-parametric regression model with the prestige
data.
}
\usage{data("Prestige")}
\format{
  A data frame with 102 observations on the following 6 variables.
  \describe{
    \item{\code{education}}{average education of occupational incumbents, years, in 1971}
    \item{\code{income}}{average income of incumbents, dollars, in 1971}
    \item{\code{women}}{percentage of incumbents who are women}
    \item{\code{prestige}}{Pineo-Porter prestige score for occupation, from a social survey conducted in the mid-
1960s}
    \item{\code{census}}{Canadian Census occupational code}
    \item{\code{type}}{a factor for type of occupation with levels \code{bc} (Blue Collar), \code{prof} (Professional, Managerial and Technical) and \code{wc} (White Collar)}
  }
}
\references{
Fox, J., Weisberg, S., Adler, D., Bates, D., Baud-Bovy, G., Ellison, S.,
Firth, D., Friendly, M., Gorjanc, G., Graves, S., Heiberger, R., Monette,
G., Murdoch, D., Nilsson, H., Ogle, D., Riply, B., Venables, W.,
Winsemius, D. and Zeileis, A. (2016). car: Companion to Applied
Regression. R package version 2.1-2.
}

