\name{snoring}
\alias{snoring}
\docType{data}
\title{Snoring and Hear Diseae Data
}
\description{
Agresti (2007) presented the data on an epidemiological survey of 2,484 subjects to investigate snoring as a possible risk factor
for heart disease. The subjects were classified according to their snoring
level, as reported by their spouses.
}
\usage{data("snoring")}
\format{
  A data frame with 4 observations on the following 3 variables.
  \describe{
    \item{\code{yes}}{number of subjects for occurrence of heart disease}
    \item{\code{no}}{number of subjects for non-occurrence of heart diease}
    \item{\code{x}}{snoring level (0, 2, 4, 5)}
  }
}
\references{
Agresti, A. (2007). An Introduction to Categorical Data Analysis. John
Wiley & Sons, New York, 2nd edition.
}

