\name{mdhglm-package}
\alias{mdhglm-package}
\docType{package}
\title{Multivariate Double Hierarchical Genearlized Linear Models}
\description{
This package allows different models for multivariate response variables where each response is assumed to follow
DHGLMs (Lee and Nelder, 2006; Lee, Nelder and Pawitan, 2006). With DHGLMs, the mean and dispersion parameters for the variance of random effects and residual variance (overdispersion) can
be modeled as random-effect models. Furthermore, the mdhglm package allows various correlation structures among the random effects in
DHGLMs of different response variables. Functions are provided to format and summarize the mdhglm results. The estimates of fixed effects,
random effects and their variances and correlations are calculated.
}

\details{
\tabular{ll}{
Package: \tab mdhglm\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2016-02-26\cr
License: \tab Unlimited\cr
LazyLoad: \tab yes\cr
}
This is version 1.2 of the mdhglm package.
}


\author{Manegseok Noh, Youngjo Lee

Maintainer: Maengseok Noh <msnoh@pknu.ac.kr>
}

\references{
Lee, Y. and Nelder, J. A. (2006). Double hierarchical generalized linear models (with discussion), Applied Statistics 55, 139--185.

Lee, Y. Nelder, J. A. and Pawitan, Y. (2006). Generalised linear models with random effects: unified analysis via
h-likelihood. Chapman & Hall: London.
}
